- Adjust distance calculation to use Manhattan distance - Generate seed data and run program on set test point (still need to enable system input through gradle) - Add tests for new code

+++ src/main/java/com/ia/dto/FillStationDistanceDTO.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 @Setter
 public class FillStationDistanceDTO {
     private CentralFillStation fillStation;
-    private double distanceFromPoint;
+    private int distanceFromPoint;
 }
